chat_project/requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.4
cryptography==38.0.3
defusedxml==0.7.1
dill==0.3.6
Django==4.1.2
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.14.0
Flask==2.2.2
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==21.3
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.3
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
